src/main/java/ru/aston/binarysearch/BinarySearch.java
package ru.aston.binarysearch;

import lombok.experimental.UtilityClass;

@UtilityClass
public class BinarySearch {

    public static int find(int[] array, int target) {
        int low = 0;
        int high = array.length - 1;

        while (low <= high) {
            int mid = (low + high) / 2;
            if (array[mid] > target) {
                high = mid - 1;
            } else if (array[mid] < target) {
                low = mid + 1;
            } else {
                return mid;
            }
        }
        return -1;
    }

}
